seperate metadata handling from key handling
