+++ lib/Controller/RequestHandlerController.php
@@ -33,6 +33,7 @@ use OCA\EndToEndEncryption\Exceptions\MetaDataExistsException;
 use OCA\EndToEndEncryption\Exceptions\MissingMetaDataException;
 use OCA\EndToEndEncryption\KeyStorage;
 use OCA\EndToEndEncryption\LockManager;
+use OCA\EndToEndEncryption\MetaDataStorage;
 use OCA\EndToEndEncryption\SignatureHandler;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\OCS\OCSBadRequestException;
@@ -62,6 +63,9 @@ class RequestHandlerController extends OCSController {
 	/** @var KeyStorage */
 	private $keyStorage;
 
+	/** @var MetaDataStorage */
+	private $metaDataStorage;
+
 	/** @var SignatureHandler */
 	private $signatureHandler;
 
@@ -84,6 +88,7 @@ class RequestHandlerController extends OCSController {
 	 * @param IRequest $request
 	 * @param string $UserId
 	 * @param KeyStorage $keyStorage
+	 * @param MetaDataStorage $metaDataStorage
 	 * @param SignatureHandler $signatureHandler
 	 * @param EncryptionManager $manager
 	 * @param LockManager $lockManager
@@ -94,6 +99,7 @@ class RequestHandlerController extends OCSController {
 								IRequest $request,
 								$UserId,
 								KeyStorage $keyStorage,
+								MetaDataStorage $metaDataStorage,
 								SignatureHandler $signatureHandler,
 								EncryptionManager $manager,
 								LockManager $lockManager,
@@ -103,6 +109,7 @@ class RequestHandlerController extends OCSController {
 		parent::__construct($AppName, $request);
 		$this->userId = $UserId;
 		$this->keyStorage = $keyStorage;
+		$this->metaDataStorage = $metaDataStorage;
 		$this->signatureHandler = $signatureHandler;
 		$this->manager = $manager;
 		$this->logger = $logger;
@@ -302,7 +309,7 @@ class RequestHandlerController extends OCSController {
 	 */
 	public function getMetaData($id) {
 		try {
-			$metaData = $this->keyStorage->getMetaData($id);
+			$metaData = $this->metaDataStorage->getMetaData($id);
 		} catch (NotFoundException $e) {
 			throw new OCSNotFoundException($this->l->t('Could not find metadata for "%s"', [$id]));
 		} catch (\Exception $e) {
@@ -327,7 +334,7 @@ class RequestHandlerController extends OCSController {
 	 */
 	public function setMetaData($id, $metaData) {
 		try {
-			$this->keyStorage->setMetaData($id, $metaData);
+			$this->metaDataStorage->setMetaData($id, $metaData);
 		} catch (MetaDataExistsException $e) {
 			return new DataResponse([], Http::STATUS_CONFLICT);
 		} catch (NotFoundException $e) {
@@ -362,7 +369,7 @@ class RequestHandlerController extends OCSController {
 		}
 
 		try {
-			$this->keyStorage->updateMetaData($id, $metaData);
+			$this->metaDataStorage->updateMetaData($id, $metaData);
 		} catch (MissingMetaDataException $e) {
 			throw new OCSNotFoundException($this->l->t("Metadata-file doesn\'t exist"));
 		} catch (NotFoundException $e) {
@@ -390,7 +397,7 @@ class RequestHandlerController extends OCSController {
 	 */
 	public function deleteMetaData($id) {
 		try {
-			$this->keyStorage->deleteMetaData($id);
+			$this->metaDataStorage->deleteMetaData($id);
 		} catch (NotFoundException $e) {
 			throw new OCSNotFoundException($this->l->t('Could not find metadata for "%s"', [$id]));
 		} catch (NotPermittedException $e) {
